@@ -1,3 +1,5 @@
+# frozen_string_literal: true
+
 class PryRails::ShowRoutes < Pry::ClassCommand
   match 'show-routes'
   group 'Rails'
@@ -8,9 +10,9 @@ class PryRails::ShowRoutes < Pry::ClassCommand
   BANNER
 
   def options(opt)
-    opt.on :G, "grep", "Filter output by regular expression",
-           :argument => true,
-           :as => Array
+    opt.on :G, 'grep', 'Filter output by regular expression',
+           argument: true,
+           as: Array
   end
 
   def process
@@ -21,11 +23,11 @@ class PryRails::ShowRoutes < Pry::ClassCommand
       if Rails::VERSION::MAJOR >= 6
         process_rails_6_and_higher(all_routes)
       elsif Rails::VERSION::MAJOR == 4 || Rails::VERSION::MAJOR == 5
-        process_rails_4_and_5(all_routes)
+        process_rails4_and5(all_routes)
       elsif Rails::VERSION::MAJOR >= 3 && Rails::VERSION::MINOR >= 2
-        process_rails_3_2(all_routes)
+        process_rails32(all_routes)
       else
-        process_rails_3_0_and_3_1(all_routes)
+        process_rails30_and31(all_routes)
       end
 
     output.puts grep_routes(formatted).join("\n")
@@ -35,6 +37,7 @@ class PryRails::ShowRoutes < Pry::ClassCommand
   # `opts[:G]`.
   def grep_routes(formatted)
     return formatted unless opts[:G]
+
     grep_opts = opts[:G]
 
     grep_opts.reduce(formatted) do |lines, pattern|
@@ -43,49 +46,49 @@ class PryRails::ShowRoutes < Pry::ClassCommand
   end
 
   # Cribbed from https://github.com/rails/rails/blob/3-1-stable/railties/lib/rails/tasks/routes.rake
-  def process_rails_3_0_and_3_1(all_routes)
+  def process_rails30_and31(all_routes)
     routes = all_routes.collect do |route|
       reqs = route.requirements.dup
       reqs[:to] = route.app unless route.app.class.name.to_s =~ /^ActionDispatch::Routing/
-        reqs = reqs.empty? ? "" : reqs.inspect
+      reqs = reqs.empty? ? '' : reqs.inspect
 
-      {:name => route.name.to_s, :verb => route.verb.to_s, :path => route.path, :reqs => reqs}
+      { name: route.name.to_s, verb: route.verb.to_s, path: route.path, reqs: reqs }
     end
 
     # Skip the route if it's internal info route
     routes.reject! { |r| r[:path] =~ %r{/rails/info/properties|^/assets} }
 
-    name_width = routes.map{ |r| r[:name].length }.max
-    verb_width = routes.map{ |r| r[:verb].length }.max
-    path_width = routes.map{ |r| r[:path].length }.max
+    name_width = routes.map { |r| r[:name].length }.max
+    verb_width = routes.map { |r| r[:verb].length }.max
+    path_width = routes.map { |r| r[:path].length }.max
 
     routes.map do |r|
       "#{r[:name].rjust(name_width)} #{r[:verb].ljust(verb_width)} #{r[:path].ljust(path_width)} #{r[:reqs]}"
     end
   end
 
-  def process_rails_3_2(all_routes)
+  def process_rails32(all_routes)
     require 'rails/application/route_inspector'
 
     Rails::Application::RouteInspector.new.format(all_routes)
   end
 
-  def process_rails_4_and_5(all_routes)
+  def process_rails4_and5(all_routes)
     require 'action_dispatch/routing/inspector'
 
-    ActionDispatch::Routing::RoutesInspector.
-      new(all_routes).
-      format(ActionDispatch::Routing::ConsoleFormatter.new).
-      split(/\n/)
+    ActionDispatch::Routing::RoutesInspector
+      .new(all_routes)
+      .format(ActionDispatch::Routing::ConsoleFormatter.new)
+      .split(/\n/)
   end
 
   def process_rails_6_and_higher(all_routes)
     require 'action_dispatch/routing/inspector'
 
-    ActionDispatch::Routing::RoutesInspector.
-      new(all_routes).
-      format(ActionDispatch::Routing::ConsoleFormatter::Sheet.new).
-      split(/\n/)
+    ActionDispatch::Routing::RoutesInspector
+      .new(all_routes)
+      .format(ActionDispatch::Routing::ConsoleFormatter::Sheet.new)
+      .split(/\n/)
   end
 
   PryRails::Commands.add_command(self)
