@@ -1,10 +1,10 @@
-# encoding: UTF-8
+# frozen_string_literal: true
 
 PryRails::Commands = Pry::CommandSet.new
 
-command_glob = File.expand_path('../commands/*.rb', __FILE__)
+command_glob = File.expand_path('commands/*.rb', __dir__)
 
-Dir[command_glob].each do |command|
+Dir[command_glob].sort.each do |command|
   require command
 end
 
