@@ -1,5 +1,5 @@
-# encoding: UTF-8
+# frozen_string_literal: true
 
 module PryRails
-  VERSION = "0.3.9"
+  VERSION = '0.3.9'
 end
