@@ -1,14 +1,13 @@
-# encoding: UTF-8
+# frozen_string_literal: true
 
 require 'spec_helper'
 
-describe "show-middleware" do
-  it "should print a list of middleware" do
+describe 'show-middleware' do
+  it 'should print a list of middleware' do
     output = mock_pry('show-middleware', 'exit-all')
 
-    output.must_match %r{^use ActionDispatch::Static$}
-    output.must_match %r{^use ActionDispatch::ShowExceptions$}
-    output.must_match %r{^run TestApp.routes\Z}
+    _(output).must_match(/^use ActionDispatch::Static$/)
+    _(output).must_match(/^use ActionDispatch::ShowExceptions$/)
+    _(output).must_match(/^run TestApp.routes\Z/)
   end
 end
-
