@@ -1,13 +1,13 @@
-# encoding: UTF-8
+# frozen_string_literal: true
 
 require 'spec_helper'
 
-describe "show-model" do
-  it "should print one ActiveRecord model" do
+describe 'show-model' do
+  it 'should print one ActiveRecord model' do
     output = mock_pry('show-model Beer', 'exit-all')
 
     expected = <<MODEL
-Beer
+\e[0GBeer
   id: integer
   name: string
   type: string
@@ -17,15 +17,15 @@ Beer
   belongs_to :hacker
 MODEL
 
-    output.must_equal expected
+    _(output).must_equal expected
   end
 
   if defined? Mongoid
-    it "should print one Mongoid model" do
+    it 'should print one Mongoid model' do
       output = mock_pry('show-model Artist', 'exit-all')
 
       expected = <<MODEL
-Artist
+\e[0GArtist
   _id: BSON::ObjectId
   name: String
   embeds_one :beer (validate)
@@ -34,22 +34,22 @@ MODEL
 
       output.gsub!(/^ *_type: String\n/, '') # mongoid 3.0 and 3.1 differ on this
       output.gsub!(/Moped::BSON/, 'BSON')    # mongoid 3 and 4 differ on this
-      output.must_equal expected
+      _(output).must_equal expected
     end
   end
 
   it "should print an error if the model doesn't exist" do
     output = mock_pry('show-model FloojBulb', 'exit-all')
-    output.must_equal "Couldn't find model FloojBulb!\n"
+    _(output).must_match(/Couldn't find model FloojBulb!/)
   end
 
   it "should print an error if it doesn't know what to do with the model" do
     output = mock_pry('show-model PryRails', 'exit-all')
-    output.must_equal "Don't know how to show PryRails!\n"
+    _(output).must_match(/Don't know how to show PryRails!/)
   end
 
-  it "should print help if no model name is given" do
+  it 'should print help if no model name is given' do
     output = mock_pry('show-model', 'exit-all')
-    output.must_match(/Usage: show-model/)
+    _(output).must_match(/Usage: show-model/)
   end
 end
