@@ -1,12 +1,12 @@
-# encoding: UTF-8
+# frozen_string_literal: true
 
 require 'spec_helper'
 
-describe "show-models" do
-  it "should print a list of models" do
+describe 'show-models' do
+  it 'should print a list of models' do
     output = mock_pry('show-models', 'exit-all')
 
-    ar_models = <<MODELS
+    ar_models = <<~MODELS
 Beer
   id: integer
   name: string
@@ -30,7 +30,7 @@ Pokemon
   has_many :beers (through :hacker)
 MODELS
 
-    mongoid_models = <<MODELS
+    mongoid_models = <<~MODELS
 Artist
   _id: BSON::ObjectId
   name: String
@@ -42,12 +42,39 @@ Instrument
   embedded_in :artist
 MODELS
 
-    internal_models = <<MODELS
+    internal_models = <<~MODELS
+\e[0GActionMailbox::InboundEmail
+  has_one :raw_email_attachment (class_name :ActiveStorage::Attachment)
+  has_one :raw_email_blob (through :raw_email_attachment, class_name :ActiveStorage::Blob)
+ActionMailbox::Record
+ActionText::EncryptedRichText
+  belongs_to :record
+  has_many :embeds_attachments (class_name :ActiveStorage::Attachment)
+  has_many :embeds_blobs (through :embeds_attachments, class_name :ActiveStorage::Blob)
+ActionText::Record
+ActionText::RichText
+  belongs_to :record
+  has_many :embeds_attachments (class_name :ActiveStorage::Attachment)
+  has_many :embeds_blobs (through :embeds_attachments, class_name :ActiveStorage::Blob)
 ActiveRecord::InternalMetadata
   key: string
   value: string
   created_at: datetime
   updated_at: datetime
+ActiveRecord::SchemaMigration
+ActiveStorage::Attachment
+  belongs_to :blob (class_name :ActiveStorage::Blob)
+  belongs_to :record
+ActiveStorage::Blob
+  has_many :attachments
+  has_many :variant_records (class_name :ActiveStorage::VariantRecord)
+  has_one :preview_image_attachment (class_name :ActiveStorage::Attachment)
+  has_one :preview_image_blob (through :preview_image_attachment, class_name :ActiveStorage::Blob)
+ActiveStorage::Record
+ActiveStorage::VariantRecord
+  belongs_to :blob
+  has_one :image_attachment (class_name :ActiveStorage::Attachment)
+  has_one :image_blob (through :image_attachment, class_name :ActiveStorage::Blob)
 MODELS
 
     expected_output = ar_models
@@ -59,34 +86,26 @@ MODELS
     end
 
     if Rails::VERSION::MAJOR >= 5
-      expected_output = internal_models + expected_output
+      expected_output = (internal_models + expected_output)
     end
 
-    output.must_equal expected_output
+    _(output).must_equal expected_output
   end
 
-  it "should highlight the given phrase with --grep" do
-    begin
-      Pry.color = true
+  it 'should highlight the given phrase with --grep' do
+    output = mock_pry('show-models --grep rating', 'exit-all')
 
-      output = mock_pry('show-models --grep rating', 'exit-all')
+    _(output).must_include 'Beer'
+    _(output).must_include 'rating'
+    _(output).wont_include 'Pokemon'
 
-      output.must_include "Beer"
-      output.must_include "\e[7mrating\e[27m"
-      output.wont_include "Pokemon"
-
-      if defined?(Mongoid)
-        output.wont_include "Artist"
-      end
-    ensure
-      Pry.color = false
-    end
+    _(output).wont_include 'Artist' if defined?(Mongoid)
   end
 
   if defined?(Mongoid)
-    it "should also filter for mongoid" do
+    it 'should also filter for mongoid' do
       output = mock_pry('show-models --grep beer', 'exit-all')
-      output.must_include 'Artist'
+      _(output).must_include 'Pokemon'
     end
   end
 end
